+++ bungee/src/main/java/com/mrivanplays/conversations/bungee/BungeeTimeoutScheduler.java
@@ -16,7 +16,7 @@ class BungeeTimeoutScheduler implements TimeoutScheduler {
   }
 
   @Override
-  public TimeoutTask schedule(Runnable task, int time, TimeUnit timeUnit) {
+  public TimeoutTask schedule(Runnable task, long time, TimeUnit timeUnit) {
     return new BungeeTimeoutTask(
         service.submit(
             () -> {
